remove parameter set generator abstraction layer
we don't need this flexibility now.
