+++ Optimizer/GridSearchOptimizationStrategy.cs
@@ -13,20 +13,82 @@
  * limitations under the License.
 */
 
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
 namespace QuantConnect.Optimizer
 {
     /// <summary>
-    /// Defines the way of enumerating all possible candidates for the solution for given optimization parameters settings
+    /// Find the best solution in first generation
     /// </summary>
-    public class GridSearch : IOptimizationParameterSetGenerator
+    public class GridSearchOptimizationStrategy : IOptimizationStrategy
     {
+        private HashSet<OptimizationParameter> _args;
+        private object _locker = new object();
+
         /// <summary>
         /// Should be global for all Step()'s
         /// </summary>
         private int _i = 0;
+
+        /// <summary>
+        /// Defines the direction of optimization, i.e. maximization or minimization
+        /// </summary>
+        public Extremum Extremum { get; private set; }
+
+        /// <summary>
+        /// Keep the best found solution - lean computed job result and corresponding  parameter set 
+        /// </summary>
+        public OptimizationResult Solution { get; private set; }
+
+        /// <summary>
+        /// Fires when new parameter set is generated
+        /// </summary>
+        public event EventHandler NewParameterSet;
+
+        /// <summary>
+        /// Initializes the strategy using generator, extremum settings and optimization parameters
+        /// </summary>
+        /// <param name="extremum">Maximize or Minimize the target value</param>
+        /// <param name="parameters">Optimization parameters</param>
+        public void Initialize(Extremum extremum, HashSet<OptimizationParameter> parameters)
+        {
+            Extremum = extremum;
+            _args = parameters;
+            _i = 0;
+        }
+
+        /// <summary>
+        /// Checks whether new lean compute job better than previous and run new iteration if necessary.
+        /// </summary>
+        /// <param name="result">Lean compute job result and corresponding parameter set</param>
+        public void PushNewResults(OptimizationResult result)
+        {
+            lock (_locker)
+            {
+                if (!ReferenceEquals(result, OptimizationResult.Empty) && result?.Target == null)
+                {
+                    // one of the requested backtests failed
+                    return;
+                }
+
+                if (result.Id > 0)
+                {
+                    if (Solution == null || Extremum.Better(Solution.Target.Value, result.Target.Value))
+                    {
+                        Solution = result;
+                    }
+
+                    return;
+                }
+
+                foreach (var parameterSet in Step(result.ParameterSet, _args))
+                {
+                    NewParameterSet?.Invoke(this, new OptimizationEventArgs(parameterSet));
+                }
+            }
+        }
 
         /// <summary>
         /// Enumerate all possible arrangements
@@ -70,5 +132,6 @@ namespace QuantConnect.Optimizer
                 }
             }
         }
+
     }
 }
