--- Optimizer/IOptimizationParameterSetGenerator.cs
@@ -1,33 +1,0 @@
-﻿/*
- * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
- * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-using System.Collections.Generic;
-
-namespace QuantConnect.Optimizer
-{
-    /// <summary>
-    /// Defines the way of generation of the new set of parameters based on optimization parameters settings
-    /// </summary>
-    public interface IOptimizationParameterSetGenerator
-    {
-        /// <summary>
-        /// Generates a new set of parameters based on optimization parameters settings
-        /// </summary>
-        /// <param name="seed">Seeding parameter set, i.e. approximation</param>
-        /// <param name="args">Optimization parameters settings</param>
-        /// <returns>New set of parameters combinations</returns>
-        IEnumerable<ParameterSet> Step(ParameterSet seed, HashSet<OptimizationParameter> args);
-    }
-}
